feat(player): add Vulkan decode fallback and overlay
Improve Vulkan playback by letting NV12 try the direct path, falling back from direct rendering failures, and exposing decoder/HW details in the player overlay.

+++ src/player/backend/ffmpeg/hw.rs
@@ -281,7 +281,7 @@ pub(super) fn vulkan_frame_planes(
         planes.truncate(raw_format.plane_count());
     }
     Ok(VulkanFrameImages {
-        usage: vk_frames.usage as u32,
+        usage: vk_frames.usage,
         planes,
     })
 }
